use int instead of size_t for some configuration values

+++ src/main/Config.cpp
@@ -450,13 +450,11 @@ Config::load(std::string const& filename)
             }
             else if (item.first == "WORKER_THREADS")
             {
-                WORKER_THREADS =
-                    static_cast<size_t>(readInt<int>(item, 1, 1000));
+                WORKER_THREADS = readInt<int>(item, 1, 1000);
             }
             else if (item.first == "MAX_CONCURRENT_SUBPROCESSES")
             {
-                MAX_CONCURRENT_SUBPROCESSES =
-                    static_cast<size_t>(readInt<int>(item, 1));
+                MAX_CONCURRENT_SUBPROCESSES = readInt<int>(item, 1);
             }
             else if (item.first == "MINIMUM_IDLE_PERCENT")
             {
